Remove Flow
When using the correct typing (according to my knowledge), Flow started
getting in the way instead of helping. Probably, it's because I was
expecting an inference as awesome as Haskell's and a piece of embedded
type information. Maybe, it's just a misuse due to knowledge lack
of how to work with Flow.